+++ src/windows.rs
@@ -347,6 +347,7 @@ impl MmapInner {
         len: usize,
         _stack: bool,
         _populate: bool,
+        _no_reserve: bool,
         _huge: Option<u8>,
     ) -> io::Result<MmapInner> {
         // Ensure a non-zero length for the underlying mapping